Refactor llvm_util
