+++ llvm_util/known_fns.h
@@ -5,6 +5,7 @@
 
 #include <memory>
 #include <vector>
+#include <functional>
 
 namespace llvm {
 class CallInst;
@@ -30,8 +31,10 @@ bool llvm_implict_attrs(llvm::Function &f, const llvm::TargetLibraryInfo &TLI,
 
 // returns true if it's a known function call
 std::pair<std::unique_ptr<IR::Instr>, bool>
-known_call(llvm::CallInst &i, const llvm::TargetLibraryInfo &TLI,
-           IR::BasicBlock &BB, const std::vector<IR::Value*> &args,
-           IR::FnAttrs &&attrs, std::vector<IR::ParamAttrs> &param_attr);
+known_call(llvm::CallInst &i, std::string &&value_name,
+           const llvm::TargetLibraryInfo &TLI, IR::BasicBlock &BB,
+           const std::vector<IR::Value*> &args, IR::FnAttrs &&attrs,
+           std::vector<IR::ParamAttrs> &param_attr,
+           std::function<IR::Value*(uint64_t val, int bits)> make_intconst);
 
 }
