+++ llvm_util/llvm2alive.cpp
@@ -6,6 +6,7 @@
 #include "llvm_util/utils.h"
 #include "util/sort.h"
 #include "llvm/ADT/SmallVector.h"
+#include "llvm/ADT/StringExtras.h"
 #include "llvm/Analysis/MemoryBuiltins.h"
 #include "llvm/Analysis/ValueTracking.h"
 #include "llvm/IR/CFG.h"
@@ -129,6 +130,11 @@ class llvm2alive_ : public llvm::InstVisitor<llvm2alive_, unique_ptr<Instr>> {
   Function *alive_fn;
   llvm::Function &f;
   const llvm::TargetLibraryInfo &TLI;
+
+  unordered_map<const llvm::Value*, Value*> value_cache;
+  unordered_map<const llvm::Value*, string> value_names;
+  unsigned value_id_counter = 0; // for %0, %1, etc..
+  
   /// True if converting a source function, false when converting a target
   /// function.
   bool IsSrc;
@@ -193,11 +199,229 @@ class llvm2alive_ : public llvm::InstVisitor<llvm2alive_, unique_ptr<Instr>> {
     return val;
   }
 
-  auto get_operand(llvm::Value *v) {
-    return llvm_util::get_operand(v,
-        [this](auto I) { return convert_constexpr(I); },
-        [this](auto ag) { return copy_inserter(ag); },
-        [this](auto fn) { return register_fn_decl(fn, nullptr); });
+  Value* make_intconst(uint64_t val, int bits) {
+    auto c = make_unique<IntConst>(get_int_type(bits), val);
+    auto ret = c.get();
+    alive_fn->addConstant(std::move(c));
+    return ret;
+  }
+
+  IR::Value* make_intconst(const llvm::APInt &val) {
+    unique_ptr<IntConst> c;
+    auto bw = val.getBitWidth();
+    auto &ty = get_int_type(bw);
+    if (bw <= 64)
+      c = make_unique<IntConst>(ty, val.getZExtValue());
+    else
+      c = make_unique<IntConst>(ty, toString(val, 10, false));
+    auto ret = c.get();
+    alive_fn->addConstant(std::move(c));
+    return ret;
+  }
+
+  IR::Value* get_poison(Type &ty) {
+    auto val = make_unique<PoisonValue>(ty);
+    auto ret = val.get();
+    alive_fn->addConstant(std::move(val));
+    return ret;
+  }
+  
+  void add_identifier(const llvm::Value &llvm, Value &v) {
+    value_cache.emplace(&llvm, &v);
+  }
+
+  void replace_identifier(const llvm::Value &llvm, Value &v) {
+    value_cache[&llvm] =  &v;
+  }
+
+  Value* get_identifier(const llvm::Value &llvm) {
+    auto I = value_cache.find(&llvm);
+    return I != value_cache.end() ? I->second : nullptr;
+  }
+
+  string value_name(const llvm::Value &v) {
+    auto &name = value_names[&v];
+    if (!name.empty())
+      return name;
+
+    if (!v.getName().empty())
+      return name = '%' + v.getName().str();
+    return name = v.getType()->isVoidTy() ? "<void>"
+                                          : "%#" + to_string(value_id_counter++);
+  }
+
+  Value* get_operand(llvm::Value *v) {
+    #define RETURN_CACHE(val)                           \
+      do {                                              \
+        auto val_cpy = val;                             \
+        ENSURE(value_cache.emplace(v, val_cpy).second); \
+        return val_cpy;                                 \
+      } while (0)
+
+    if (auto ptr = get_identifier(*v))
+      return ptr;
+
+    auto ty = llvm_type2alive(v->getType());
+    if (!ty)
+      return nullptr;
+
+    if (auto cnst = dyn_cast<llvm::ConstantInt>(v)) {
+      RETURN_CACHE(make_intconst(cnst->getValue()));
+    }
+
+    if (auto cnst = dyn_cast<llvm::ConstantFP>(v)) {
+      auto &apfloat = cnst->getValueAPF();
+      auto c
+      = make_unique<FloatConst>(*ty,
+                                toString(apfloat.bitcastToAPInt(), 10, false),
+                                true);
+      auto ret = c.get();
+      alive_fn->addConstant(std::move(c));
+      RETURN_CACHE(ret);
+    }
+
+    if (isa<llvm::PoisonValue>(v)) {
+      RETURN_CACHE(get_poison(*ty));
+    }
+
+    if (isa<llvm::UndefValue>(v)) {
+      auto val = make_unique<UndefValue>(*ty);
+      auto ret = val.get();
+      alive_fn->addUndef(std::move(val));
+      RETURN_CACHE(ret);
+    }
+
+    if (isa<llvm::ConstantPointerNull>(v)) {
+      auto val = make_unique<NullPointerValue>(*ty);
+      auto ret = val.get();
+      alive_fn->addConstant(std::move(val));
+      RETURN_CACHE(ret);
+    }
+
+    if (auto gv = dyn_cast<llvm::GlobalVariable>(v)) {
+      if (hasOpaqueType(gv->getValueType()))
+        // TODO: Global variable of opaque type is not supported.
+        return nullptr;
+
+      unsigned size = DL().getTypeAllocSize(gv->getValueType());
+      unsigned align = gv->getPointerAlignment(DL()).value();
+      string name;
+      if (!gv->hasName()) {
+        unsigned id = 0;
+        auto M = gv->getParent();
+        auto i = M->global_begin(), e = M->global_end();
+        for (; i != e; ++i) {
+          if (i->hasName())
+            continue;
+          if (&(*i) == gv)
+            break;
+          ++id;
+        }
+        assert(i != e);
+        name = '@' + to_string(id);
+      } else {
+        name = '@' + gv->getName().str();
+      }
+
+      bool arb_size = false;
+      if (auto *arr = dyn_cast<llvm::ArrayType>(gv->getValueType()))
+        arb_size = arr->getNumElements() == 0;
+
+      auto val = make_unique<GlobalVariable>(*ty, std::move(name), size, align,
+                                            gv->isConstant(), arb_size);
+      auto gvar = val.get();
+      alive_fn->addConstant(std::move(val));
+      RETURN_CACHE(gvar);
+    }
+
+    if (auto fn = dyn_cast<llvm::Function>(v)) {
+      auto val = make_unique<GlobalVariable>(
+        *ty, '@' + fn->getName().str(), 0,
+        fn->getAlign().value_or(llvm::Align(8)).value(), true, true);
+      auto gvar = val.get();
+      alive_fn->addConstant(std::move(val));
+
+      if (!register_fn_decl(fn, nullptr))
+        return nullptr;
+
+      RETURN_CACHE(gvar);
+    }
+
+    auto fillAggregateValues = [&](AggregateType *aty,
+        function<llvm::Value *(unsigned)> get_elem, vector<Value*> &vals) -> bool
+    {
+      unsigned opi = 0;
+
+      for (unsigned i = 0; i < aty->numElementsConst(); ++i) {
+        if (!aty->isPadding(i)) {
+          if (auto op = get_operand(get_elem(opi)))
+            vals.emplace_back(op);
+          else
+            return false;
+          ++opi;
+        }
+      }
+      return true;
+    };
+
+    if (auto cnst = dyn_cast<llvm::ConstantAggregate>(v)) {
+      vector<Value*> vals;
+      if (!fillAggregateValues(dynamic_cast<AggregateType *>(ty),
+              [&cnst](auto i) { return cnst->getOperand(i); }, vals))
+        return nullptr;
+
+      auto val = make_unique<AggregateValue>(*ty, std::move(vals));
+      auto ret = val.get();
+      if (all_of(cnst->op_begin(), cnst->op_end(), [](auto &V) -> bool
+          { return isa<llvm::ConstantData>(V); })) {
+        alive_fn->addConstant(std::move(val));
+        RETURN_CACHE(ret);
+      } else {
+        alive_fn->addAggregate(std::move(val));
+        return copy_inserter(ret);
+      }
+    }
+
+    if (auto cnst = dyn_cast<llvm::ConstantDataSequential>(v)) {
+      vector<Value*> vals;
+      if (!fillAggregateValues(dynamic_cast<AggregateType *>(ty),
+              [&cnst](auto i) { return cnst->getElementAsConstant(i); }, vals))
+        return nullptr;
+
+      auto val = make_unique<AggregateValue>(*ty, std::move(vals));
+      auto ret = val.get();
+      alive_fn->addConstant(std::move(val));
+      RETURN_CACHE(ret);
+    }
+
+    if (auto cnst = dyn_cast<llvm::ConstantAggregateZero>(v)) {
+      vector<Value*> vals;
+      if (!fillAggregateValues(dynamic_cast<AggregateType *>(ty),
+              [&cnst](auto i) { return cnst->getElementValue(i); }, vals))
+        return nullptr;
+
+      auto val = make_unique<AggregateValue>(*ty, std::move(vals));
+      auto ret = val.get();
+      alive_fn->addConstant(std::move(val));
+      RETURN_CACHE(ret);
+    }
+
+    if (auto cexpr = dyn_cast<llvm::ConstantExpr>(v)) {
+      return convert_constexpr(cexpr);
+    }
+
+    // This must be an operand of an unreachable instruction as the operand
+    // hasn't been seen yet
+    if (isa<llvm::Instruction>(v)) {
+      auto val = make_unique<PoisonValue>(*ty);
+      auto ret = val.get();
+      alive_fn->addConstant(std::move(val));
+      return ret;
+    }
+
+    return nullptr;
+
+    #undef RETURN_CACHE
   }
 
   RetTy NOP(llvm::Instruction &i) {
@@ -217,7 +441,6 @@ public:
         out(&get_outs()) {}
 
   ~llvm2alive_() {
-    reset_state();
     for (auto &inst : i_constexprs) {
       inst->deleteValue();
     }
@@ -403,8 +626,11 @@ public:
     }
 
     if (!approx) {
-      auto [known, approx0]
-        = known_call(i, TLI, *BB, args, std::move(attrs), param_attrs);
+      auto [known, approx0] = known_call(i, value_name(i), TLI, *BB, args,
+                                         std::move(attrs), param_attrs,
+                                         [&](uint64_t val, int bits) -> Value* {
+                                          return make_intconst(val, bits);
+                                         });
       approx = approx0;
       if (known)
         return std::move(known);
@@ -1790,7 +2016,6 @@ public:
     Function Fn(*type, f.getName().str(), 8 * DL().getPointerSize(),
                 DL().getIndexSizeInBits(0), DL().isLittleEndian(),
                 f.isVarArg());
-    reset_state(Fn);
 
     alive_fn = &Fn;
     BB = &Fn.getBB("#init", true);
